update exponential primitives
