+++ featuretools/primitives/standard/__init__.py
@@ -2,7 +2,6 @@
 from featuretools.primitives.base.aggregation_primitive_base import AggregationPrimitive
 from featuretools.primitives.standard.aggregation_primitives import *
 from featuretools.primitives.standard.datetime_transform_primitives import *
-from featuretools.primitives.standard.exponential_transform_primitives import *
 from featuretools.primitives.standard.latlong_transform_primitives import *
 from featuretools.primitives.standard.rolling_transform_primitives import *
 from featuretools.primitives.standard.transform_primitive import *
