+++ featuretools/primitives/standard/transform_primitives/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 from featuretools.primitives.standard.transform_primitives.binary_transform_primitives import *
 from featuretools.primitives.standard.transform_primitives.cumulative_transform_primitives import *
+from featuretools.primitives.standard.transform_primitives.exponential_transform_primitives import *
 from featuretools.primitives.standard.transform_primitives.natural_language_primitives import *
