+++ featuretools/primitives/standard/transform_primitives/exponential_transform_primitives/__init__.py
@@ -1,0 +1,9 @@
+from featuretools.primitives.standard.transform_primitives.exponential_transform_primitives.exponential_weighted_average import (
+    ExponentialWeightedAverage,
+)
+from featuretools.primitives.standard.transform_primitives.exponential_transform_primitives.exponential_weighted_std import (
+    ExponentialWeightedSTD,
+)
+from featuretools.primitives.standard.transform_primitives.exponential_transform_primitives.exponential_weighted_variance import (
+    ExponentialWeightedVariance,
+)
